requirements.txt
langchain==0.0.184
python-dotenv==1.0.0
openai==0.27.6
faiss-cpu==1.7.4
altair==4
tiktoken==0.4.0
flask==2.3.2
scikit-learn==1.3.0
transformers==4.32.0
flask_cors==4.0.0